add SSD_TEST for read
