+++ SSD_Excutor/SSD_Excutor.vcxproj
@@ -150,6 +150,7 @@
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">   %(AdditionalOptions)</AdditionalOptions>
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Release|x64'">   %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
+    <ClCompile Include="SSDHelper.cpp" />
     <ClCompile Include="ssd_test.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">  %(AdditionalOptions)</AdditionalOptions>
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">  %(AdditionalOptions)</AdditionalOptions>
