+++ SSD_Excutor/SSD_Excutor.vcxproj.filters
@@ -51,6 +51,9 @@
     <ClCompile Include="ssd_test.cpp">
       <Filter>테스트파일</Filter>
     </ClCompile>
+    <ClCompile Include="SSDHelper.cpp">
+      <Filter>테스트파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
